API: Rutas de backend creadas respecto a nuevas tablas
